Add support for grepping for the pattern only at certain fields instead of the whole line.
Add support for searching the patterns only at certain fields instead of
the whole line. To limit the searching of the patterns to certain field
a comma separated list can be given to the -f option.

The default field separator is set to TAB, but can be changed with the
-s option.

The error message are now also written to stderr instead of stdout.

Signed-off-by: Gert Hulselmans <hulselmansgert@gmail.com>