Make pattern matching much faster for -w option when multiple patterns need to be checked.
The pattern matching is now much faster for the -w option by doing a key lookup
in the grep_pattern_array array.

The grep_pattern_array now doesn't use integer numbers as keys but it uses
the patterns as keys. This allows to lookup fields/lines as a key in the
grep_pattern_array array. Key lookup in arrays is much faster than looping
over each element of the array in awk and comparing the value stored for
that key with a field/line.

Signed-off-by: Gert Hulselmans <hulselmansgert@gmail.com>